clean up components/headerstory
