+++ pages/_app.tsx
@@ -1,7 +1,7 @@
 import "../styles/globals.css";
 import "../styles/mapcss.css";
 import type { AppProps } from "next/app";
-import Header from "../components/Header";
+import Header from "../components/Header/Header";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
